+++ fuel-tx/src/builder.rs
@@ -235,7 +235,8 @@ impl TransactionBuilder<Blob> {
 impl TransactionBuilder<Mint> {
     pub fn mint(
         block_height: BlockHeight,
-        tx_index: u16,
+        #[cfg(feature = "u32-tx-pointer")] tx_index: u32,
+        #[cfg(not(feature = "u32-tx-pointer"))] tx_index: u16,
         input_contract: input::contract::Contract,
         output_contract: output::contract::Contract,
         mint_amount: Word,